backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.0
PyMuPDF==1.25.1
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
tomli==2.1.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.18.3
